move access tests into a separate file
